// frontend/src/app/core/api/model/exam-time-slot-response.ts
/**
 * Training school system Backend API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ExamTimeSlotResponse { 
    id?: number;
    code?: string;
    label?: string;
    startTime?: string;
    endTime?: string;
    createdDate?: string;
    modifiedDate?: string;
    isActive?: boolean;
}